staff-service/requirements.txt
Django==4.2.27
djangorestframework==3.15.2
requests==2.32.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
graphene-django==3.2.2
django-cors-headers==4.6.0
